// ev-backoffice/client/src/app/task-queue/models/package-modal-type.enum.ts
export enum PackageModalType {
  IMMIGRATION_CATEGORY_CONFLICT = 'IMMIGRATION_CATEGORY_CONFLICT',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  BLOCKED_OPEN_PACKAGES = 'BLOCKED_OPEN_PACKAGES',
  PAYMENT_CHARGED = 'PAYMENT_CHARGED',
  MEMBERS_OF_BLOCKED_PACKAGE = 'MEMBERS_OF_BLOCKED_PACKAGE',
  MEMBERS_WITH_PENDING_OR_DENY_STATUS = 'MEMBERS_WITH_PENDING_OR_DENY_STATUS',
  REMINDER_APPLICANT_INVITATION = 'REMINDER_APPLICANT_INVITATION',
  REMINDER_APPLICANT_PERMISSION = 'REMINDER_APPLICANT_PERMISSION',
}